core: move `CalculateNetworkFee` to a separate package
