+++ pkg/rpc/server/client_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/nspcc-dev/neo-go/pkg/config/netmode"
-	"github.com/nspcc-dev/neo-go/pkg/core"
+	"github.com/nspcc-dev/neo-go/pkg/core/fee"
 	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/crypto/hash"
 	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
@@ -105,7 +105,7 @@ func TestAddNetworkFee(t *testing.T) {
 		}}
 		require.NoError(t, c.AddNetworkFee(tx, 10, accs[0]))
 		require.NoError(t, accs[0].SignTx(tx))
-		cFee, _ := core.CalculateNetworkFee(accs[0].Contract.Script)
+		cFee, _ := fee.Calculate(accs[0].Contract.Script)
 		require.Equal(t, int64(io.GetVarSize(tx))*feePerByte+cFee+10, tx.NetworkFee)
 	})
 
@@ -129,8 +129,8 @@ func TestAddNetworkFee(t *testing.T) {
 		require.NoError(t, accs[0].SignTx(tx))
 		require.NoError(t, accs[1].SignTx(tx))
 		require.NoError(t, accs[2].SignTx(tx))
-		cFee, _ := core.CalculateNetworkFee(accs[0].Contract.Script)
-		cFeeM, _ := core.CalculateNetworkFee(accs[1].Contract.Script)
+		cFee, _ := fee.Calculate(accs[0].Contract.Script)
+		cFeeM, _ := fee.Calculate(accs[1].Contract.Script)
 		require.Equal(t, int64(io.GetVarSize(tx))*feePerByte+cFee+cFeeM+10, tx.NetworkFee)
 	})
 	t.Run("Contract", func(t *testing.T) {
